+++ test/OpenTelemetry.Extensions.Benchmarks/Program.cs
@@ -1,0 +1,7 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+using BenchmarkDotNet.Running;
+
+var summaries = BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+return summaries.SelectMany(p => p.Reports).Any((p) => !p.Success) ? 1 : 0;